Use Float64 to prevent overflow when squaring Int64
(to compute standard deviation)

+++ src/show.jl
@@ -149,9 +149,9 @@ function _print_timer(io::IO, to::TimerOutput, ∑t::Integer, ∑b::Integer, ind
     maxt = accum_data.maxtime
     nc = accum_data.ncalls
     t2 = accum_data.timesq
-    avgt = t/nc
+    avgt = Float64(t)/nc
     avgt2 = t2/nc
-    stdt = sqrt((avgt2 - avgt^2)*nc/(nc-1))
+    stdt = sqrt((avgt2-avgt^2)*nc/(nc-1))
 
     name = truncdots(to.name, name_length - indent)
     print(io, " ")
